requirements.txt
langchain==0.1.0
langchain-google-genai==1.0.1
langchain-community==0.0.13
python-dotenv==1.0.0
requests==2.31.0
PyGithub==1.59.1
pydantic==2.5.3
pydantic-settings==2.1.0
fastapi==0.108.0
uvicorn==0.25.0
python-multipart==0.0.6
schedule==1.2.0
aiohttp==3.9.1
pandas==2.1.4
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.0
jinja2==3.1.2
